updated new version of mp-spdz, wip

+++ GC/Clear.h
@@ -17,11 +17,11 @@ public:
     static string type_string() { return "clear"; }
 
     Clear() : Integer() {}
-    Clear(long a) : Integer(a) {}
+    Clear(long long a) : Integer(a) {}
     Clear(const IntBase& x) { IntBase::operator=(x); }
 
     void xor_(const Clear& x, const Clear& y) { a = x.a ^ y.a; }
-    void xor_(const Clear& x, long y) { a = x.a ^ y; }
+    void xor_(const Clear& x, long long y) { a = x.a ^ y; }
 };
 
 } /* namespace GC */
